+++ __tests__/01_basic_spec.ts
@@ -1,6 +1,4 @@
 import {
-  BrowserRouter,
-  Route,
   LazyFetcher,
   useRouteData,
   useRouteDataSelector,
@@ -9,8 +7,6 @@ import {
 
 describe('basic spec', () => {
   it('exported function', () => {
-    expect(BrowserRouter).toBeDefined();
-    expect(Route).toBeDefined();
     expect(LazyFetcher).toBeDefined();
     expect(useRouteData).toBeDefined();
     expect(useRouteDataSelector).toBeDefined();