+++ examples/03_nesting/src/App.tsx
@@ -3,14 +3,14 @@ import React, { Suspense } from 'react';
 import { BrowserRouter } from 'react-suspense-router';
 
 import Nav from './Nav';
-import Routes from './Routes';
+import MyRoutes from './MyRoutes';
 
 const App: React.FC = () => (
-  <BrowserRouter suspenseConfig={{ timeoutMs: 1000 }}>
+  <BrowserRouter timeout={1000}>
     <Nav />
     <hr />
     <Suspense fallback={<div>Loading...</div>}>
-      <Routes />
+      <MyRoutes />
     </Suspense>
   </BrowserRouter>
 );