+++ src/index.ts
@@ -1,44 +1,78 @@
 export {
-  MemoryRouter,
-  Prompt,
+  MemoryRouter, // TODO
+  Navigate,
+  Outlet,
   Redirect,
+  Route,
   Router,
-  StaticRouter, // TODO?
-  Switch,
+  createRoutesFromChildren,
   generatePath,
-  matchPath,
-  withRouter,
-  useHistory,
+  matchRoutes,
+  resolveLocation,
+  useBlocker,
+  useHref,
   useLocation,
+  useMatch,
+  useNavigate,
+  useOutlet,
   useParams,
-  useRouteMatch,
-  match, // just a type
+  useResolvedLocation,
+  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+  // @ts-ignore
 } from 'react-router';
 
 export {
   HashRouter, // TODO
   Link,
   NavLink,
+  Prompt,
+  usePrompt,
+  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+  // @ts-ignore
 } from 'react-router-dom';
 
 /**
- * BrowserRouter for Suspsense
+ * BrowserRouter for Suspsense Render-as-You-Fetch
  *
- * This accepts `suspenseConfig` prop for useTransition.
+ * Its usage is the same with react-router.
  *
  * @example
  * import { BrowserRouter } from 'react-suspense-router';
  *
  * const App = () => (
- *   <BrowserRouter suspenseConfig={{ timeoutMs: 3000 }}>
+ *   <BrowserRouter timeoutMs={3000}>
  *     <Nav />
  *     <Suspense fallback={<span>Loading...</span>}>
- *       <Routes />
+ *       <MyRoutes />
  *     </Suspense>
  *   </BrowserRouter>
  * );
  */
 export { BrowserRouter } from './BrowserRouter';
+
+/**
+ * useRoutes for Suspense Render-as-You-Fetch
+ *
+ * Its usage is the same with react-router.
+ */
+export { useRoutes } from './Routes';
+
+/**
+ * Routes for Suspense Render-as-You-Fetch
+ *
+ * Its usage is the same with react-router.
+ *
+ * @example
+ * import { Routes, Route } from 'react-suspense-router';
+ *
+ * const MyRoutes = () => (
+ *   <Routes>
+ *     <Route path="/" element={<Index />} />
+ *     <Route path="/double/:number" element={<Double />} fetchData={fetchDouble} />
+ *   </Routes>
+ * );
+ */
+export { Routes } from './Routes';
 
 /**
  * LazyFetcher
@@ -51,28 +85,6 @@ export { BrowserRouter } from './BrowserRouter';
  * const fetchUserData = LazyFetcher(() => import('./pages/User.data'));
  */
 export { LazyFetcher } from './LazyFetcher';
-
-/**
- * Route for Suspense
- *
- * This accepts `fetchData` prop for route data.
- * It is usaually loaded by `LazyFetcher`.
- *
- * @example
- * import { Route } from 'react-suspense-router';
- *
- * const Routes = () => (
- *   <>
- *     <Route exact path="/">
- *       <Index />
- *     </Route>
- *     <Route exact path="/user/:uid" fetchData={fetchUserData}>
- *       <User />
- *     </Route>
- *   </>
- * );
- */
-export { Route } from './Route';
 
 /**
  * useRouteData hook
@@ -127,3 +139,6 @@ export { useRouteDataSelector } from './RouteDataContext';
  * };
  */
 export { useSuspensePending } from './SuspensePendingContext';
+
+// types
+export { match } from './types';